only logged in users can delete and edit articles now

+++ article.php
@@ -2,6 +2,9 @@
 
 require 'classes/Database.php';
 require 'classes/Article.php';
+require 'classes/Auth.php';
+
+session_start();
 
 $db = new Database();
 $conn = $db->getConn();
@@ -23,8 +26,13 @@ if (isset($_GET['id'])) {
     <h2><?= htmlspecialchars($article->title); ?></h2>
     <p><?= htmlspecialchars($article->content); ?></p>
   </article>
-  <a href="edit-article.php?id=<?= $article->id; ?>">Edit</a>
-  <a href="delete-article.php?id=<?= $article->id; ?>">Delete</a>
+
+  <?php if (Auth::isLoggedIn()) : ?>
+    <a href="edit-article.php?id=<?= $article->id; ?>">Edit</a>
+    <a href="delete-article.php?id=<?= $article->id; ?>">Delete</a>
+  <?php else : ?>
+    <p></p>
+  <?php endif; ?>
 
 <?php else : ?>
